requirements.txt
Flask==3.1.1
python-dotenv==1.1.1
google-generativeai==0.8.5
requests==2.32.4